Basic UI commit
Done with basic UI
