+++ app/src/main/java/com/manish/grihasthi/UpdateProduct.java
@@ -9,5 +9,6 @@ public class UpdateProduct extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_update_product);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
     }
 }
