+++ app/src/main/res/drawable/gradient_add_form.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item>
+        <shape>
+            <gradient
+
+                android:startColor="#ffbfc3"
+                android:centerColor="#f3bfff"
+                android:endColor="#6875ec"
+                android:angle="90"
+                />
+
+        </shape>
+
+    </item>
+
+</selector>
