--- app/src/main/res/layout/activity_add_update_product.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".UpdateProduct"
-    android:background="@color/AddUpdateProduct">
-
-<android.support.v7.widget.LinearLayoutCompat
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <android.support.design.widget.TextInputLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginRight="20px"
-        android:layout_marginLeft="20px">
-
-        <EditText
-            android:id="@+id/p_name"
-            android:layout_width="match_parent"
-            android:layout_height="40dp"
-            android:layout_marginBottom="10dp"
-            android:hint="Product Name"
-            android:inputType="textPersonName" />
-    </android.support.design.widget.TextInputLayout>
-    <android.support.design.widget.TextInputLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginRight="20px"
-        android:layout_marginLeft="20px">
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/p_price"
-            android:hint="Product Price"
-            android:inputType=""
-            android:layout_marginBottom="10dp"/>
-    </android.support.design.widget.TextInputLayout>
-    <android.support.design.widget.TextInputLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginRight="20px"
-        android:layout_marginLeft="20px">
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/p_stock"
-            android:hint="Product Name"
-            android:inputType="textPersonName"
-            android:layout_marginBottom="10dp"/>
-    </android.support.design.widget.TextInputLayout>
-    <android.support.design.widget.TextInputLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginRight="20px"
-        android:layout_marginLeft="20px">
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/p_quantity"
-            android:hint="Product Name"
-            android:inputType="textPersonName"
-            android:layout_marginBottom="10dp"/>
-    </android.support.design.widget.TextInputLayout>
-
-
-</android.support.v7.widget.LinearLayoutCompat>
-</RelativeLayout>
