+++ app/src/main/res/menu/activity_main_page_drawer.xml
@@ -6,15 +6,21 @@
     <group android:checkableBehavior="single">
         <item
             android:id="@+id/add_product_nav"
-            android:icon="@drawable/add_product_icon"
+            android:icon="@drawable/icons8_new_64"
             android:title="Add Product" />
         <item
             android:id="@+id/update_product_nav"
-            android:icon="@drawable/update_product_icon"
+            android:icon="@drawable/icons8_video_editing_64"
             android:title="Update Product" />
         <item
+            android:id="@+id/add_agency_nav"
+            android:icon="@drawable/icons8_supplier_64"
+            android:title="Add Agency" />
+
+
+        <item
             android:id="@+id/add_company_nav"
-            android:icon="@drawable/add_company_icon"
+            android:icon="@drawable/icons8_commodity_64"
             android:title="Add Company" />
 
     </group>
@@ -23,11 +29,11 @@
         <menu>
             <item
                 android:id="@+id/company_stock_nav"
-                android:icon="@drawable/company_wise_logo"
+                android:icon="@drawable/icons8_gold_bars_64"
                 android:title="By Company" />
             <item
                 android:id="@+id/product_stock_nav"
-                android:icon="@drawable/product_wise_logo"
+                android:icon="@drawable/icons8_product_64"
                 android:title="By Product" />
         </menu>
     </item>
